pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitops-replacer"
version = "0.1.0"
authors = [
    {name = "Simon Lauger", email = "simon@lauger.de"}
]
description = "Automated YAML/JSON value updates for GitOps repositories via GitHub API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["gitops", "yaml", "json", "kubernetes", "helm", "automation", "github", "ci-cd", "deployment"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "requests>=2.25.0",
    "PyYAML>=5.4.0",
]

[project.urls]
Homepage = "https://github.com/slauger/gitops-replacer"
Documentation = "https://github.com/slauger/gitops-replacer#readme"
Repository = "https://github.com/slauger/gitops-replacer"
Issues = "https://github.com/slauger/gitops-replacer/issues"
Changelog = "https://github.com/slauger/gitops-replacer/blob/main/CHANGELOG.md"

[project.scripts]
gitops-replacer = "gitops_replacer.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
build_command = "pip install build && python -m build"
commit_parser = "angular"
major_on_zero = true
tag_format = "v{version}"

[tool.semantic_release.commit_message]
message = "chore(release): {version} [skip ci]"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease = false
